+++ clang/docs/ReleaseNotes.rst
@@ -473,6 +473,9 @@ Improvements to Clang's diagnostics
        }
      };
 
+- Clang now diagnoses missing format attributes for non-template functions and
+  class/struct/union members. Fixes #GH70024
+
 Improvements to Clang's time-trace
 ----------------------------------